Create attestation model to remove circular dependency
